Cleanup linkTo and getMinimalAction
